Add "MikMod for Rockbox 0.1" from 2007-06-29
This is an early port of MikMod to Rockbox that I did back in 2007.
It was one of my early exercises in the C language.

Since I didn't upstream it back then, someone else ended up porting it
with a nicer UI.

Fun fact:
This version's PCM buffers are designed to be just about large enough
to allow an iPod Video to render fm-2year/rapido.xm without pausing.